requirements.txt
playwright==1.40.0
pytest==7.4.3
pytest-playwright==0.4.3
pytest-xdist==3.3.1
pytest-html==4.1.1
allure-pytest==2.13.2
pytest-rerunfailures==13.0
pytest-timeout==2.2.0
python-dotenv==1.0.0
requests==2.31.0
faker==20.0.0
pydantic==2.4.2
typing-extensions==4.8.0 